requirements.txt
py-cord[voice]
PyNaCl

# discord.py>=2.3.1
gTTS>=2.3.2
python-dotenv>=1.0.0
requests>=2.31.0